Scene Generator now supports any object
You can now generate the entirety of the level using any object (or objects!) you like. The spray pattern and patterns in general have been improved. A preview (with a color modification attached to it) has been included to show the width and height more easily. Lastly this window recieved the ability to name the levels.
The main window has also been reordenated.